Allow duplicating complication on long press
